Enum relocation + initial renaming
