+++ src/main/java/frc/robot/commands/AlgaeScoreSequence.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 //import frc.robot.commands.SetElevator.SetElevatorCancelBehavior;
 import frc.robot.subsystems.AlgaeMechanism;
 import frc.robot.subsystems.Elevator;
-import frc.robot.subsystems.Elevator.ElevatorPosition;
+import static frc.robot.Enums.ElevatorEnums.*;
 
 public class AlgaeScoreSequence extends SequentialCommandGroup {
     public AlgaeScoreSequence(AlgaeMechanism algaeMechanism, Elevator elevator, ElevatorPosition cryingPosition) {
