+++ src/main/java/frc/robot/commands/CoralIntakeSequence.java
@@ -6,14 +6,11 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WrapperCommand;
-import frc.robot.commands.MoveCoralMechanism.CoralMechanismPosition;
-import frc.robot.commands.MoveCoralMechanism.MoveCoralCancelBehavior;
-import frc.robot.commands.RunCoralIntake.CoralIntakeDirection;
 import frc.robot.commands.SetElevator.SetElevatorCancelBehavior;
 import frc.robot.subsystems.CoralMechanism;
 import frc.robot.subsystems.Elevator;
-import frc.robot.subsystems.Elevator.ElevatorPosition;
-import frc.robot.subsystems.Elevator.LogicalElevatorPosition;
+import static frc.robot.Enums.ElevatorEnums.*;
+import static frc.robot.Enums.CoralEnums.*;
 
 /**
  * Moves the coral mechanism to the intake position, and runs the intake to suck coral in.
@@ -33,7 +30,7 @@ public class CoralIntakeSequence extends WrapperCommand {
                     new ConditionalCommand(
                         new SetElevator(elevator, ElevatorPosition.CORAL_INTAKE, SetElevatorCancelBehavior.CANCEL_SETPOINT_REACHED),
                         new InstantCommand(),
-                        () -> elevator.getLogicalElevatorPosition() == LogicalElevatorPosition.L1),
+                        () -> elevator.getCurrentTarget() == TargetLevel.L1),
                     new MoveCoralMechanism(coralMechanism,
                                            CoralMechanismPosition.INTAKE,
                                            MoveCoralCancelBehavior.CANCEL_SETPOINT_REACHED)),
